Initial phase of factoring out maxtry settings from ddc_try_stats.c into ddc_retry_mgt.c. - max tries are now thread specific - max try settings still to be removed from ddc_try_stats.c,   currently croos-checked wih asserts for now
ddc_try_stats.c/h:
- Try_Data struct no longer referenced as an opaque pointer
  in function calls.  Made code pointlessly obscure.
- Try_Data struct now contains a DDCA_Retry_Type value.
  makes usage clearer and simpler
- try_data_create() now takes an additional  DDCA_Retry_Type argument

ddc_multi_part_io.c, ddc_packet_io.c:
- change try_data_create() calls as per above

api_base.c:
- ddca_get_max_tries(), ddca_set_max_tries():
  - use ddc_retry_mgt.c function ddc_get_thread_single_max_try()

parsed_cmd.h:
- change max_tries type to uint16_t

cmd_parser_goption.c:
- parse maxtries values into unsigned value

main.c:
- include ddc_retry_mgt.h
- comments

+++ src/cmdline/parsed_cmd.h
@@ -79,7 +79,7 @@ struct {
    gchar **            traced_files;
    gchar **            traced_functions;
    DDCA_Output_Level   output_level;
-   int                 max_tries[3];
+   uint16_t            max_tries[3];
    float               sleep_multiplier;
    uint32_t            flags;      // Parsed_Cmd_Flags
    int                 i1;         // available for temporary use
